break out some tests
